DAC6-3722 | update a move view tests from controller spec to a view spec
