+++ app/views/UnauthorisedStandardUserView.scala.html
@@ -43,9 +43,9 @@ showBackLink = true) {
 
     @para(Html(messages("unauthorisedStandardUser.paragraph2")),classes = "govuk-body govuk-!-margin-bottom-1")
 
-    @para(Html(messages("unauthorisedStandardUser.paragraph3")))
+    @para(Html(messages("unauthorisedStandardUser.link")))
 
-    @para(Html(messages("unauthorisedStandardUser.paragraph4")))
+    @para(Html(messages("unauthorisedStandardUser.paragraph3")))
 
 
 }
